agents/sequential_orchestrator_flat/pyproject.toml
[project]
name = "sequential-orchestrator-flat"
version = "0.1.0"
description = "扁平化实现的写作和评分流程编排器"
authors = [
    {name = "ADK团队"}
]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-adk = ">=0.0.2"
google-genai = "^1.9.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
google-adk = { version = ">=0.0.2", extras = ["eval"] }
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api" 